Adding support Object.prototype.toString.call
**Changes**
 * Added unit tests for new test pattern
 * Provided commented example of how it should work in unit tests
  * Ensured that it matched coding style of nearby code
 * Added '%C' placeholder support for Class type.

**Notes**
Most of the code is written in ES5 or below style. I have modified the
changes to follow this style of programming. Tests detect capabilities
before making assumptions about the test environment, so in places where
Symbols do not exist, for example, no attempt is made to unit test them.